pyproject.toml
[project]
name = "finananceragagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.8",
    "langchain-core>=0.3.34",
    "langchain-groq>=0.2.4",
    "langchain-mongodb>=0.4.0",
    "langgraph>=0.2.70",
    "langgraph-checkpoint-mongodb>=0.1.0",
    "opik>=1.4.11",
    "pre-commit>=4.1.0",
    "pydantic-settings>=2.7.1",
    "pymongo>=4.9.2",
    "loguru>=0.7.3",
    "langchain-huggingface>=0.1.2",
    "langchain-community>=0.3.17",
    "ipykernel>=6.29.5",
    "pydantic>=2.10.6",
    "datasketch>=1.6.5",
    "langchain-openai>=0.3.17",
    "langchain-experimental>=0.3.4",
    "pandas>=2.2.3",
]
